Update Guessing gaming github mainfile.py
Guessing Game with simple and basic concepts :) Here is a guessing game with simple and basic concepts where while, if, break and simple conditional statements are used. In this game you will type the number you guessed and computer will also guess the number, if the number guessed by you and the computer is same then you will geta point. So this is the brief information about the game. PLAY AND ENJOY THE GAME :)